src/pyproject.toml
[tool.poetry]
name = "pymap"
version = "2.0.0-alpha"
description = ""
authors = ["hallowslab <hallow.tech.lab@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
celery = {extras = ["redis"], version = "^5.3.6"}
flower = "^2.0.1"
django-celery-beat = {git = "https://github.com/celery/django-celery-beat.git"}
taskipy = "^1.8.0"
docutils = "^0.20.1"
redis = {extras = ["hiredis"], version = "^5.0.1"}

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
pytest = "^6.2.5"
autoflake = "^1.4"
coverage = "^6.3.2"
django-debug-toolbar = "^4.2.0"
mypy = ">=1.7.0"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}

[tool.taskipy.tasks]
serverDev = "DJANGO_ENV='development' task server"
workerDev = "DJANGO_ENV='development' task worker"
monitorDev = "DJANGO_ENV='development' task monitor"
server = "python manage.py runserver 0.0.0.0:8000"
worker = "celery -A pymap.celery worker"
monitor = "celery -A pymap.celery flower --basic-auth='admin:admin,dev:localhost' --address='localhost'"
### Database tasks
upgradeDev = "DJANGO_ENV='development' task upgrade"
upgrade = "task prep && task migr"
prep = "python manage.py makemigrations"
migr = "python manage.py migrate"
### Formatting, tests
check = "DJANGO_ENV='development' task format"
format = "task black && task flake && task mypy"
black = "black . "
flake = "autoflake -i -r ."
mypy = "mypy --strict ."
testCore = "pytest tests/core"
### Coverage
coverage = "coverage run -m pytest && coverage report"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
